refactor old tally test
